rpc: Add client::compat::discover
A utility function to discover the compatibility mode using a
pre-instantiated RPC client. This has a bit of a chicken-and-egg
problem, but it should work in practice as long as it's the first
operation done for the client.

HttpClient gets a set_compat_mode method to update the mode
on the fly. No such thing is provided for WebSocketClient, though,
because the compat mode is set into the subscription driver
and cannot be updated.