rpc: support more changes in CometBFT 0.38
Add the new fields and types without static differentiation for 0.38.x.

Use the domain type ExecTxResult in deserializing the endpoint response types in the current (i.e. 0.37+) dialect for these endpoints:
- /block_results
- /broadcast_tx_commit
- /tx
- /tx_commit

The Deserialize impls are restored on the response types,
corresponding to the JSON schema for the current RPC dialect.

Simplify the 0.34 dialect support for these responses, no longer using
generics.

In endpoint::tx_commit::Response, rename the deliver_tx field to
tx_result to correspond to the current JSON schema. The Deserialize
impl falls back to deliver_tx to also be able to parse 0.37 responses.