Support for CometBFT proto 0.38 (#1312)
* Generate protos for v0_38

Check out the protos from the v0.38.0-alpha.1 release tag and
add the v0_38 fork of generated modules to tendermint-proto.

* proto: add serializers for v0_38::types::Evidence

* p2p: switch tendermint-proto imports to v0_38

* tendermint: domain type for AbciParams

The AbciParams message/object was added in 0.38, used as the
"abci" field in ConsensusParams.
Add protobuf conversions for v0_38 protos, and serde.

* tendermint: BlockIdFlag as a domain type

This is needed for the more detailed VoteInfo in 0.38.

* Extend VoteInfo to support 0.38 data

The BlockSignatureInfo enum is an attempt to represent both
the pre-0.38 encoding with the signed_last_block flag, and the
current one where the signature information is given by the
block_id_flag field.

* Rename Signature::to_bytes

Rename to Signature::into_bytes to follow Rust naming conventions.

* tendermint: v0_38 serialization for Vote

The serialization of extension fields is not tested yet.

* proto: use v0_38 in BlockId unit tests

* tendermint: v0_38 serialization for block::Size

* tendermint: v0_38 conversions for abci::Event

* Extend tendermint_pb_modules to v0_38

* abci: Update to ABCI 0.38

Change the tendermint-proto structs in use to v0_38
and update the application API accordingly.

* light-client-verifier: adapt to new Vote fields

* CI: update buf and protoc to latest versions

* CI: diff between generated and checked-in files

* Restore serde derives on PublicKey

There was no apparent value added by defining the impls manually,
IDK why the attributes were previously shifted to the enum type.

* Regenerate protos from latest release tags

Use v0.34.28 and v0.37.1

* tendermint: v0_38 proto support for request::Info

* tendermint: support 0.38 extensions for VoteInfo

Add members to VoteInfo to be able to convert from and to
the ExtendedVoteInfo message type as defined in 0.38 proto.
Define conversions from and to the proto type, as well as
conversions between CommitInfo and ExtendedCommitInfo in the
proto.

As a helper to the above, provide Bytes conversions for the
Signature domain type.

* tendermint: v0.38 ABCI requests and responses

Add the v0_38::abci::{Request,Response} enums and the new domain types
representing request and response messages added in CometBFT 0.38.

* tendermint: version-specific kinded response enums

Define {ConsensusRequest, InfoRequest, MempoolRequest, SnapshotRequest}
separately for each of the v0_*::abci versioned modules, enumerating
just the requests that can be used with the respective version.
No panic.

* abci: crude logic and tests for finalize_block

* tendermint: Re-export v0_38 kinded ABCI enums

Just like with abci::Request and abci::Response, the categorized
request and response enums get re-exported from the v0_38::abci
module.

* tendermint: domain type for ExtendedVoteInfo

The ExtendedVoteInfo and ExtendedCommitInfo protobuf messages
can be encoded only in RequestPrepareProposal. In other contexts
the extension fields are not useful.