RPC support for CometBFT 0.38 (#1317)
* rpc: 0.38 support for /block_results endpoint

In the v0_37+ dialect helper, support the new finalize_block_events
field if it is present. Also default the begin_block_events and
end_block_events fields to None if they are not present.

* rpc: add app_hash field to /block_results response

* tendermint: restore Deserialize on some types

For abci::Event and abci::response::CheckTx, restore the Deserialize
implementation on domain types themselves.

The event serialization format as used in 0.37+ will be established as
the "canonical" serialization.

* tendermint: serde impls for ExecTxResult

* rpc: support more changes in CometBFT 0.38

Add the new fields and types without static differentiation for 0.38.x.

Use the domain type ExecTxResult in deserializing the endpoint response types in the current (i.e. 0.37+) dialect for these endpoints:
- /block_results
- /broadcast_tx_commit
- /tx
- /tx_commit

The Deserialize impls are restored on the response types,
corresponding to the JSON schema for the current RPC dialect.

Simplify the 0.34 dialect support for these responses, no longer using
generics.

In endpoint::tx_commit::Response, rename the deliver_tx field to
tx_result to correspond to the current JSON schema. The Deserialize
impl falls back to deliver_tx to also be able to parse 0.37 responses.

* rpc: recognize 0.38 in CompatMode::from_version

* Fix deserialization of RPC results

Some serde attributes were not migrated onto the ExecTxResult
domain type when it has been tasked with deserializing the latest
RPC dialect.
On abci::response::CheckTx, add serde(default) on fields that are
no longer present in CometBFT 0.38.

* rpc: 0.37 compat on /block_results response

If the finalize_block_events is not present, default to an empty vector.
Conversely, skip the field from serializing if it's set to an
empty vector.

* Add kvstore fixtures for 0.38

* Derive serde impls for abci responses

BeginBlock, EndBlock, FinalizeBlock are used to deserialize
RPC responses.

* rpc: adapt serialization of Event for 0.38

Now that event data can have new fields, distinguish this with
a new NewBlock enum variant, renaming the old one to LegacyNewBlock.

The dialect serialization helpers for subscription events are
separated to v0_34::DialectEvent, that does not recognize the
0.38 fields at all and parses the event attributes as base64-encoded,
and latest::DialectEvent, which has provisions for 0.38 fields if they
are present, and in this case converts to the NewBlock variant.

* base64 for app_hash in JSON of FinalizeBlock

* rpc: version-alias latest event serde helpers

For consistency, define event::v0_37 and event::v0_38, aliasing
the public definitions in event::latest.

* rpc: split ser/de helpers for Event

The Serialize impl needs to be implemented differently for
0.37 and 0.38 in order to emit old or new fields for NewBlock.
The Deserialize impl, however, flexibly handles both formats to avoid
introducing another dialect where most other helpers would be
identical. This means that Event gets two different helpers for each
protocol version: DeEvent for Deserialize and SerEvent for Serialize.

* rpc: 0.37 serialization for /broadcast_tx_commit

Provide v0_37::DialectResponse as a way to serialize responses
in the 0.37 JSON format.