Validate construction and deserialization of `validator::Set` (#1350)
* proto: deserialize total_voting_power in ValidatorSet

If the field is present, deserialize it into the struct field
rather than skipping it completely. This will allow validation
when deserializing validator::Set in tendermint.

* tendermint: harden total_voting_power computation for Set

Calculate the total voting power without overflows and validate
against the protocol-defined limit, provided as the
Set::MAX_TOTAL_VOTING_POWER const.

In the conversion from tendermint-proto struct VaidatorSet,
check the unmarshalled total_voting_power value to match the calculated
sum of the list's validators, unless it's the default 0.

* tendermint: deserialize validator::Set via TryFrom

Use the validating conversion from the protobuf struct
ValidatorSet, which should have the same schema including
the optionally deserialized total_voting_power field.
This is to ensure we don't blindly trust total_voting_power if it's
present in the serialization, and accept when it's absent.

* tendermint: fix formatting of InvalidKey error

The Display impl called itself in infinite recursion.

* proto: restore custom PublicKey serializers

These are used in serialization for ValidatorSet and
LightClientAttackEvidence.

* proto: add helper module serializers::from_str_allow_null

* proto: allow Validator.proposer_priority be null

Some light-client-js tests use fixtures where it was null.
The field used to be skipped in previous serialization.