p2p: replace `x25519-dalek` with `curve25519-dalek-ng`
The `tendermint-p2p` crate previously pulled in both an obsolete version
of `x25519-dalek`/`curve25519-dalek`, but also `curve25519-dalek-ng` by
way of `ed25519-consensus`.

This replaces `x25519-dalek` with `curve25519-dalek-ng` which is already
a dependency of `ed25519-consensus`. This reduces the overall number of
dependencies and ensures `curve25519-dalek-ng` is the main Curve25519
implementation used.