[WIP] p2p: add `AsyncSecretConnection` based on tokio
Adds an `async` version of `SecretConnection`, currently specialized to
TCP-based connections exclusively.

This requires quite a bit of duplication, especially since the existing
`SecretConnection` implementation is entirely built around `std::io`
traits, which are unusable in an async context.

The implementation returns whole chunks as a `Vec<u8>`, which avoids the
need to "double buffer" incoming data as ownership of such data is
passed directly to the caller upon completing a read.